% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.center.R
\name{l.center}
\alias{l.center}
\title{Center a Level-1 Variable}
\usage{
l.center(data, user_id, var_id, type = "pmc")
}
\arguments{
\item{data}{dataset stored as dataframe}

\item{user_id}{the name of the level-2 grouping variable (e.g., person id); can be stored as a factor, numeric, or integer within a dataframe}

\item{var_id}{the name of the level-1 variable you want to center; should be stored as numeric or integer within a dataframe}

\item{type}{the type of centering you want to use, either person-mean centering \code{pmc} (i.e., centering around the means of level-2 grouping variables) or grand-mean centering \code{gmc} (i.e., centering around the grand mean); defaults to \code{pmc}; see Hoffman & Stawksi (2009) for a discussion of centering approaches}
}
\value{
a dataframe with a centered variable appended to it; if using person-mean centering, the person-mean variable will also be appended
}
\description{
For centering level-1 variables, either with person-mean (i.e., within-person) or grand-mean centering. Using the function will output a new dataset with both the centered and group mean variables.
}
\examples{
new_ds <- l.center(data = ds, user_id = 'id', var_id = 'l1_var', type = 'pmc')
}
\references{
Hoffman, L., & Stawski, R. S. (2009). Persons as contexts: Evaluating between-person and within-person effects in longitudinal analysis. \emph{Research in Human Development, 6}(2-3), 97-120.
}
\seealso{
\code{\link{l.describe}}
}
\author{
Myles A. Maillet, \email{myles.a.maillet@gmail.com}
}
\keyword{centering}
\keyword{variables,}
\keyword{within-person}
