% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.describe.R
\name{l.describe}
\alias{l.describe}
\title{Calculate Descriptive Statistics for a Level-1 Variable}
\usage{
l.describe(data, user_id, var_id, time_id = NULL, dec = 2)
}
\arguments{
\item{data}{dataset stored as dataframe}

\item{user_id}{the name of the level-2 grouping variable (e.g., person id)}

\item{var_id}{the name of the level-1 variable you want to examine; should be stored as numeric or integer within a dataframe}

\item{time_id}{if longitudinal, the name of the level-1 time variable (i.e., measurement occasion); should be stored as numeric or integer within a dataframe; if ignored, descriptive statistics are calculated by person id}

\item{dec}{the number of decimals used to display descriptive statistics; defaults to 2}
}
\value{
returns a matrix with descriptive statistics for each time point or user
}
\description{
Calculates several descriptive statistics for a level-1 variable. Can be grouped by time or person id (i.e., level-2 grouping variable).
}
\examples{
l.describe(data = ds, user_id = 'id', var_id = 'l1_var')
l.describe(data = ds, user_id = 'id', var_id = 'l1_var', time_id = 'time')
}
\author{
Myles A. Maillet, \email{myles.a.maillet@gmail.com}
}
\keyword{describe}
\keyword{descriptive}
\keyword{statistics,}
