% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.icc.R
\name{l.icc}
\alias{l.icc}
\title{Calculate the ICC for a Level-1 Variable}
\usage{
l.icc(
  data,
  user_id,
  var_id,
  var_type = "continuous",
  time_id = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{dataset stored as dataframe}

\item{user_id}{the name of the level-2 grouping variable (i.e., person id); can be stored as a factor, numeric, or integer within a dataframe}

\item{var_id}{the name of the level-1 variable you want to calculate an ICC for; should be stored as numeric or integer within a dataframe}

\item{var_type}{the type of level-1 variable you are using, either \code{continuous} or \code{binary}; defaults to \code{continuous}; see Austin & Merlo (2017) for details on how the ICC is calculated with a binary outcome variable}

\item{time_id}{if included, will calculate the ICC using an unconditional growth model (i.e., time included in the model as a predictor); should be stored as numeric or integer within a dataframe}

\item{verbose}{if \code{TRUE}, outputs a summary statement along with the ICC; defaults to \code{FALSE}}
}
\value{
returns a value for the ICC
}
\description{
Calculates an ICC for a level-1 variable by fitting a two-level unconditional model or unconditional growth model to determine the proportion of variance at each level. The ICC reflects the proportion of variance that is attributed to the higher level (e.g., ICC = 0.40 indicates that 40\% of the variance in a variable is at level-2).
}
\examples{
l.icc(data = ds, user_id = 'id', var_id = 'l1_var')
l.icc(data = ds, user_id = 'id', var_id = 'l1_var', time_id = 'time')
l.icc(data = ds, user_id = 'id', var_id = 'l1_var', var_type = 'binary', verbose = TRUE)
}
\references{
Austin, P.C. & Merlo, J. (2017). Intermediate and advanced topics in multilevel logistic regression analysis. \emph{Statistics in medicine, 36}(20), 3257-3277.
}
\seealso{
\code{\link{l.describe}}
}
\author{
Myles A. Maillet, \email{myles.a.maillet@gmail.com}
}
\keyword{coefficient}
\keyword{correlation}
\keyword{icc,}
\keyword{intraclass}
