% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.missbyid.R
\name{l.missbyid}
\alias{l.missbyid}
\title{Examine Missing Data for a Variable by User}
\usage{
l.missbyid(data, user_id, time_id, missing = "count")
}
\arguments{
\item{data}{dataset stored as dataframe; missingness for all variables will be printed, so it is recommended to only include variables of interest in the dataset}

\item{user_id}{the name of the level-2 grouping variable (e.g., person id)}

\item{time_id}{the name of the \strong{time} variable (i.e., measurement occasion); should be stored as numeric or integer within a dataframe}

\item{missing}{missing data presented as either a \code{count} or \code{percent}; defaults to \code{count}}
}
\value{
returns a matrix summary of the number (or percent) of missing observations by person id
}
\description{
Tabulates a matrix summary of the number (or percent) or missing data for each variable in a dataset for each participant. Assumes all participants are expected to provide data for each time point (i.e., measurement occasion). Requires missing data to be stored as \code{NA}.
}
\examples{
new_ds <- l.missbyid(data = ds, user_id = 'id', time_id = 'time', missing = 'count')
}
\seealso{
\code{\link{l.missbytime}}

\code{\link{l.naconvert}}
}
\author{
Myles A. Maillet, \email{myles.a.maillet@gmail.com}
}
\keyword{data}
\keyword{missing}
