% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.missfill.R
\name{l.missfill}
\alias{l.missfill}
\title{Impute NA Values for Missing Observations in a Longitudinal Design}
\usage{
l.missfill(data, user_id, time_id)
}
\arguments{
\item{data}{dataset stored as dataframe}

\item{user_id}{the name of the level-2 grouping variable (e.g., person id); can be stored as a factor, numeric, or integer within a dataframe}

\item{time_id}{if included, will calculate the ICC using an unconditional growth model (i.e., time included in the model as a predictor); should be stored as numeric or integer within a dataframe}
}
\value{
returns an imputed dataset with \code{NA} values for missing observations
}
\description{
Imputes NA values for missing (but expected) observations for each participant at each time point in a longitudinal design. In other words, creates a dataframe that includes all possible measurement occasions for each user. Limited use as-is, but a dependency for other missing data functions in this package. Requires missing data to be stored as \code{NA}.
}
\examples{
new_ds <- l.missfill(data = ds, user_id = 'id', time_id = 'time')
}
\seealso{
\code{\link{l.missbytime}}

\code{\link{l.missbyid}}

\code{\link{l.naconvert}}
}
\author{
Myles A. Maillet, \email{myles.a.maillet@gmail.com}
}
\keyword{data,}
\keyword{imputation}
\keyword{missing}
