% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.naconvert.R
\name{l.naconvert}
\alias{l.naconvert}
\title{Convert Missing Data Values to NA}
\usage{
l.naconvert(data, miss_value)
}
\arguments{
\item{data}{dataset stored as dataframe}

\item{miss_value}{current value specifying missing data (e.g., -999, \code{NULL})}
}
\value{
returns a dataframe with missing data values converted to \code{NA}
}
\description{
This function converts a value (e.g., -999) to NA, which is the required for the use of other functions in this package that involve missing data.
}
\examples{
new_ds <- l.naconvert(data = ds, miss_value = -999)
}
\author{
Myles A. Maillet, \email{myles.a.maillet@gmail.com}
}
\keyword{data}
\keyword{missing}
